requirements.txt
asgiref==3.3.1
cloudinary==1.25.0
Django==3.1.6
django-heroku==0.3.1
django-bootstrap-pagination==1.7.1
django-ckeditor==6.0.0
django-cleanup==4.0.0
django-cloudinary-storage==0.3.0
django-crispy-forms==1.11.0
django-elasticsearch-dsl==7.1.4
django-excel==0.0.10
django-extensions==3.1.0
django-formtools==2.2
django-import-export==2.5.0
django-js-asset==1.2.2
django-mptt==0.11.0
django-mptt-admin==2.0.2
django-phone-field==1.8.1
django-phonenumber-field==5.0.0
django-widget-tweaks==1.4.8
phonenumbers==8.12.17
Pillow==7.0.0
pytz==2019.3
sqlparse==0.3.0
gunicorn==20.1.0
whitenoise==5.2.0